put comments and remove useless file
